refactor: 폴더 구조 변경 및 필요없는 파일 제거
- 어플리케이션에 필요없는 테스트 파일들 제거
