+++ omg-react-app/src/App.js
@@ -1,19 +1,15 @@
-import './App.css';
-import Home from "./Home";
+import Home from "./pages/Home";
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import Main from "./pages/Main";
 function App() {
   return (
     <Router>
       <Routes>
         <Route path='/' element={<Home />} />
-        <Route path='/main' element={<MainPage />} />
+        <Route path='/main' element={<Main />} />
       </Routes>
     </Router>
   );
-}
-
-function MainPage() {
-  return <h1>Main Page</h1>
 }
 
 export default App;
