+++ omg-react-app/src/pages/Main.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+
+const Main = () => {
+    return (
+        <h1>메인 페이지입니다.</h1>
+    );
+};
+
+export default Main;
